+++ src/StkTable/StkTable.vue
@@ -32,6 +32,7 @@
         :style="{
             '--row-height': props.autoRowHeight ? void 0 : virtualScroll.rowHeight + 'px',
             '--header-row-height': props.headerRowHeight + 'px',
+            '--footer-row-height': props.footerRowHeight + 'px',
             '--highlight-duration': props.highlightConfig.duration && props.highlightConfig.duration + 's',
             '--highlight-timing-function': highlightSteps ? `steps(${highlightSteps})` : void 0,
             '--sb-width': `${scrollbarOptions.width}px`,
@@ -101,7 +102,7 @@
                     </tr>
                 </thead>
 
-                <tfoot v-if="footData && footData.length > 0" style="position: sticky; bottom: 0; left: 0; z-index: 1">
+                <tfoot v-if="footData && footData.length > 0">
                     <tr v-for="(footRow, footRowIndex) in footData" :key="footRowIndex">
                         <td v-if="virtualX_on" class="vt-x-left"></td>
                         <td v-for="(col, colIndex) in virtualX_columnPart" :key="colKeyGen(col)" v-bind="getTFProps(col)">
@@ -119,7 +120,7 @@
                                 {{ footRow[col.dataIndex] ?? getEmptyCellText(col, footRow) }}
                             </div>
                             <div v-else-if="col.type === 'seq'" class="table-cell-wrapper">
-                                {{ footRow[col.dataIndex] ?? '' }}
+                                {{ footRow[col.dataIndex] ?? 'Total' }}
                             </div>
                             <div v-else class="table-cell-wrapper" :title="footRow[col.dataIndex] || ''">
                                 <span v-if="footRow[col.dataIndex] != null">{{ footRow[col.dataIndex] }}</span>
@@ -357,6 +358,8 @@ const props = withDefaults(
         rowCurrentRevokable?: boolean;
         /** 表头行高。default = rowHeight */
         headerRowHeight?: number | string | null;
+        /** 表尾行高。default = rowHeight */
+        footerRowHeight?: number | string | null;
         /** 虚拟滚动 */
         virtual?: boolean;
         /** x轴虚拟滚动(必须设置列宽)*/
@@ -489,6 +492,7 @@ const props = withDefaults(
         rowActive: () => DEFAULT_ROW_ACTIVE_CONFIG,
         rowCurrentRevokable: true,
         headerRowHeight: DEFAULT_ROW_HEIGHT,
+        footerRowHeight: DEFAULT_ROW_HEIGHT,
         virtual: false,
         virtualX: false,
         columns: () => [],
@@ -1171,7 +1175,12 @@ function getTFProps(col: StkTableColumn<DT>) {
     return {
         'data-col-key': colKey,
         style: cellStyleMap.value[TagType.TF].get(colKey),
-        class: [col.className, fixedColClassMap.value.get(colKey), col.align === 'center' ? 'text-c' : col.align === 'right' ? 'text-r' : ''],
+        class: [
+            col.className,
+            fixedColClassMap.value.get(colKey),
+            col.type === 'seq' ? 'seq-column' : '',
+            col.align === 'center' ? 'text-c' : col.align === 'right' ? 'text-r' : '',
+        ],
     };
 }